requirements.txt
arabic-reshaper==2.1.4
asgiref==3.6.0
asn1crypto==1.5.1
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.11.2
boto3==1.26.64
botocore==1.29.64
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cryptography==38.0.4
cssselect2==0.7.0
cycler==0.11.0
decorator==5.1.1
Django==4.1.4
django-crispy-forms==1.14.0
django-storages==1.13.2
executing==1.2.0
fonttools==4.38.0
future==0.18.2
google-play-scraper==1.2.3
gunicorn==20.1.0
html5lib==1.1
idna==3.4
ipython==8.10.0
japanize-matplotlib==1.1.3
jedi==0.18.2
jmespath==1.0.1
joblib==1.2.0
kaleido==0.2.1
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.6.2
matplotlib-inline==0.1.6
mecab-python3==1.0.6
networkx==3.0
nlplot==1.6.0
numpy==1.24.1
oscrypto==1.3.0
packaging==22.0
pandas==1.5.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
plotly==5.13.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pyHanko==0.16.0
pyhanko-certvalidator==0.19.8
pyparsing==3.0.9
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
s3transfer==0.6.0
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.1
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
stack-data==0.6.2
svglib==1.4.1
tenacity==8.2.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tqdm==4.64.1
traitlets==5.9.0
tzdata==2022.7
tzlocal==4.2
unidic-lite==1.0.8
uritools==4.0.0
urllib3==1.26.13
wcwidth==0.2.6
webencodings==0.5.1
wordcloud==1.8.2.2
xhtml2pdf==0.2.8
